Moving the DbConnectionClosed to Common
